# docs/swagger.yaml
basePath: /
definitions:
...
  dto.UserLoginRequest:
    properties:
      password:
        description: Password 密码
        type: string
      username:
        description: Username 用户名
        type: string
    required:
    - password
    - username
    type: object
  dto.UserRegisterRequest:
    properties:
      password:
        description: Password 密码
        type: string
      username:
        description: Username 用户名
        type: string
    required:
    - password
    - username
    type: object
...
  vo.UserVO:
    properties:
      id:
        description: ID 用户ID
        type: integer
      username:
        description: Username 用户名
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API支持
    url: http://www.swagger.io/support
  description: 这是一个用户排名管理系统的API文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 排名系统API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: 用户登录，传入用户名和密码
      parameters:
      - description: 登录信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户登录
      tags:
      - user
...
  /api/register:
    post:
      consumes:
      - application/json
      description: 用户注册，传入用户名和密码
      parameters:
      - description: 注册信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/vo.UserVO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 用户注册
      tags:
      - user
...
schemes:
- http
swagger: "2.0"